Centralize adding sentiment related records
